+++ app/config/security_dev.yml
@@ -1,27 +1,27 @@
 # you can read more about security in the related section of the documentation
 # http://symfony.com/doc/current/book/security.html
-#security:
-#    # http://symfony.com/doc/current/book/security.html#where-do-users-come-from-user-providers
-#    providers:
-#
-#        graviton_airlock_authentication_key_user_provider:
-#            id: graviton_airlock_authentication_key_user_provider
-#
-#    # the main part of the security, where you can set up firewalls
-#    # for specific sections of your app
-#    firewalls:
-#        # disables authentication for assets and the profiler, adapt it according to your needs
-#        dev:
-#            pattern:  ^/(_(profiler|wdt)|css|images|js)/
-#            security: false
-#
-#        # graviton access restriction
-#        secured_area:
-#            #  no pattern means matches all routes
-#            stateless: true
-#            simple_preauth:
-#                authenticator: graviton_airlock_authentication_key_authenticator
-#
+security:
+    # http://symfony.com/doc/current/book/security.html#where-do-users-come-from-user-providers
+    providers:
+
+        graviton_airlock_authentication_key_user_provider:
+            id: graviton_airlock_authentication_key_user_provider
+
+    # the main part of the security, where you can set up firewalls
+    # for specific sections of your app
+    firewalls:
+        # disables authentication for assets and the profiler, adapt it according to your needs
+        dev:
+            pattern:  ^/(_(profiler|wdt)|css|images|js)/
+            security: false
+
+        # graviton access restriction
+        secured_area:
+            #  no pattern means matches all routes
+            stateless: true
+            simple_preauth:
+                authenticator: graviton_airlock_authentication_key_authenticator
+
 #    # with these settings you can restrict or allow access for different parts
 #    # of your application based on roles, ip, host or methods
 #    # http://symfony.com/doc/current/cookbook/security/access_control.html
