+++ src/Graviton/SecurityBundle/Authentication/AirlockAuthenticationKeyAuthenticator.php
@@ -16,18 +16,13 @@ use Symfony\Component\Security\Core\Authentication\Token\PreAuthenticatedToken;
 use Symfony\Component\Security\Core\Authentication\Token\TokenInterface;
 use Symfony\Component\Security\Core\Exception\AuthenticationException;
 use Symfony\Component\Security\Core\User\UserProviderInterface;
-use Symfony\Component\Security\Http\Authentication\AuthenticationFailureHandlerInterface;
-use Symfony\Component\Security\Http\Authentication\AuthenticationSuccessHandlerInterface;
 
 /**
  * @author   List of contributors <https://github.com/libgraviton/graviton/graphs/contributors>
  * @license  http://opensource.org/licenses/gpl-license.php GNU Public License
  * @link     http://swisscom.ch
  */
-final class AirlockAuthenticationKeyAuthenticator implements
-    SimplePreAuthInterface,
-    AuthenticationFailureHandlerInterface,
-    AuthenticationSuccessHandlerInterface
+final class AirlockAuthenticationKeyAuthenticator implements SimplePreAuthInterface
 {
     /**
      * @var \Graviton\SecurityBundle\User\AirlockAuthenticationKeyUserProvider
@@ -117,51 +112,5 @@ final class AirlockAuthenticationKeyAuthenticator implements
     public function supportsToken(TokenInterface $token, $providerKey)
     {
         return $token instanceof PreAuthenticatedToken && $token->getProviderKey() === $providerKey;
-    }
-
-    /**
-     * This is called when an interactive authentication attempt fails. This is
-     * called by authentication listeners inheriting from
-     * AbstractAuthenticationListener.
-     *
-     * @param Request                 $request   original request
-     * @param AuthenticationException $exception exception from auth attempt
-     *
-     * @return Response|null
-     */
-    public function onAuthenticationFailure(Request $request, AuthenticationException $exception)
-    {
-        $this->logger->warning(
-            $exception->getMessageKey(),
-            array(
-                'data' => $exception->getMessageData(),
-            )
-        );
-
-        return new Response(
-            $exception->getMessageKey(),
-            Response::HTTP_NETWORK_AUTHENTICATION_REQUIRED
-        );
-    }
-
-    /**
-     * This is called when an interactive authentication attempt succeeds. This
-     * is called by authentication listeners inheriting from
-     * AbstractAuthenticationListener.
-     *
-     * @param Request        $request Current request to be processed
-     * @param TokenInterface $token   Current token containing the authentication information
-     *
-     * @return Response|null
-     */
-    public function onAuthenticationSuccess(Request $request, TokenInterface $token)
-    {
-        $this->logger->info(
-            sprintf(
-                'Contract (%s (%s)) was successfully recognized.',
-                $token->getUsername(),
-                $token->getUser()->getContractNumber()
-            )
-        );
     }
 }
