+++ src/Graviton/SecurityBundle/Tests/Authentication/AirlockAuthenticationKeyAuthenticatorTest.php
@@ -193,72 +193,6 @@ class AirlockAuthenticationKeyAuthenticatorTest extends \PHPUnit_Framework_TestC
     }
 
     /**
-     * @return void
-     */
-    public function testOnAuthenticationFailure()
-    {
-        $exceptionMock = $this->getMockBuilder('\Symfony\Component\Security\Core\Exception\AuthenticationException')
-            ->disableOriginalConstructor()
-            ->setMethods(array('getMessageKey'))
-            ->getMock();
-        $exceptionMock
-            ->expects($this->exactly(2))
-            ->method('getMessageKey')
-            ->will($this->returnValue('test_message'));
-
-        $authenticator = new AirlockAuthenticationKeyAuthenticator(
-            $this->getProviderMock(),
-            $this->getStrategyMock(),
-            $this->logger
-        );
-
-        $response = $authenticator->onAuthenticationFailure(new Request(), $exceptionMock);
-
-        $this->assertEquals('test_message', $response->getContent());
-        $this->assertEquals(511, $response->getStatusCode());
-    }
-
-    /**
-     * @return void
-     */
-    public function testOnAuthenticationSuccess()
-    {
-        $this->logger
-            ->expects($this->once())
-            ->method('info')
-            ->with($this->equalTo('Contract (Jon Doe (1234567)) was successfully recognized.'));
-
-        $authenticator = new AirlockAuthenticationKeyAuthenticator(
-            $this->getProviderMock(),
-            $this->getStrategyMock(),
-            $this->logger
-        );
-
-        $userDouble = $this->getMockBuilder('\Graviton\SecurityBundle\Entities\SecurityContract')
-            ->disableOriginalConstructor()
-            ->setMethods(array('getContractNumber'))
-            ->getMock();
-        $userDouble
-            ->expects($this->once())
-            ->method('getContractNumber')
-            ->will($this->returnValue('1234567'));
-
-        $tokenDouble = $this->getMockBuilder('\Symfony\Component\Security\Core\Authentication\Token\TokenInterface')
-            ->setMethods(array('getUsername', 'getUser'))
-            ->getMockForAbstractClass();
-        $tokenDouble
-            ->expects($this->once())
-            ->method('getUsername')
-            ->will($this->returnValue('Jon Doe'));
-        $tokenDouble
-            ->expects($this->once())
-            ->method('getUser')
-            ->will($this->returnValue($userDouble));
-
-        $authenticator->onAuthenticationSuccess(new Request(), $tokenDouble);
-    }
-
-    /**
      * @param string[] $methods methods to mock
      *
      * @return \PHPUnit_Framework_MockObject_MockObject|AirlockAuthenticationKeyUserProvider
